+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/fluentassert/verify/compare/v1.1.0...HEAD)
 
+### Added
+
+- Add `FailureMessage.Err` method together with `AssertionError` type
+  to represent assertion results as `error` type.
+
 ## [1.1.0](https://github.com/fluentassert/verify/releases/tag/v1.1.0) - 2024-02-06
 
 This release adds length assertions.